[SPARK-56803][SQL][FOLLOWUP] Fix NPE in DowncastLongUpdater.decodeSingleDictionaryId
`DowncastLongUpdater` targets a 32-bit decimal column vector (precision <= 9),
which stores values in `intData[]`; `longData[]` is unallocated. The pre-existing
`decodeSingleDictionaryId` called `values.putLong(...)`, which NPE'd as soon as
the dictionary-decode path was exercised by the end-to-end vectorized read test
added in SPARK-56803. Narrow the dictionary's long value to int with the same
`(int) longValue` cast already used by `readValue` and `readValues`.

+++ sql/core/src/main/java/org/apache/spark/sql/execution/datasources/parquet/ParquetVectorUpdaterFactory.java
@@ -706,7 +706,13 @@ public class ParquetVectorUpdaterFactory {
         WritableColumnVector values,
         WritableColumnVector dictionaryIds,
         Dictionary dictionary) {
-      values.putLong(offset, dictionary.decodeToLong(dictionaryIds.getDictId(offset)));
+      // Target is a 32-bit decimal column vector (precision <= 9); narrow the dictionary's
+      // long value to int via the same `(int) longValue` cast used by `readValue` and
+      // `readValues`. The pre-existing `putLong` here was a latent bug that did not surface
+      // until the dictionary-decode path was exercised by the new end-to-end test, because
+      // a 32-bit decimal vector stores values in `intData` and `putLong` would NPE on the
+      // unallocated `longData`.
+      values.putInt(offset, (int) dictionary.decodeToLong(dictionaryIds.getDictId(offset)));
     }
   }
